feature oriented architecture for Speaker domain
