+++ config/routes.yaml
@@ -5,10 +5,6 @@ index:
 fos_js_routing:
     resource: "@FOSJsRoutingBundle/Resources/config/routing/routing.xml"
 
-speaker:
-    resource: routes/speakers.yaml
-    prefix: speakers
-
 sponsorship_level:
     resource: routes/sponsorshipLevel.yaml
     prefix: sponsorship-level
