+++ src/Speaker/Index/IndexAction.php
@@ -2,17 +2,16 @@
 
 declare(strict_types=1);
 
-namespace App\Controller\Speaker;
+namespace App\Speaker\Index;
 
-use App\Repository\SpeakerRepositoryInterface;
-use Ramsey\Uuid\Uuid;
+use App\Action;
+use App\Speaker\SpeakerRepositoryInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Twig\Environment as Twig;
 
-final class Show
+final class IndexAction implements Action
 {
     private $renderer;
     private $speakerRepository;
@@ -28,15 +27,10 @@ final class Show
      */
     public function handle(Request $request): Response
     {
-        $id = Uuid::fromString($request->attributes->get('id'))->toString();
+        $speakers = $this->speakerRepository->findAll();
 
-        $speaker = $this->speakerRepository->find($id);
-        if (!$speaker) {
-            throw new NotFoundHttpException();
-        }
-
-        return new Response($this->renderer->render('speaker/show.html.twig', [
-            'speaker' => $speaker,
+        return new Response($this->renderer->render('speaker/index.html.twig', [
+            'speakers' => $speakers,
         ]));
     }
 }
