+++ src/Speaker/Show/ShowAction.php
@@ -1,0 +1,43 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Speaker\Show;
+
+use App\Action;
+use App\Speaker\SpeakerRepositoryInterface;
+use Ramsey\Uuid\Uuid;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
+use Twig\Environment as Twig;
+
+final class ShowAction implements Action
+{
+    private $renderer;
+    private $speakerRepository;
+
+    public function __construct(Twig $renderer, SpeakerRepositoryInterface $speakerRepository)
+    {
+        $this->renderer = $renderer;
+        $this->speakerRepository = $speakerRepository;
+    }
+
+    /**
+     * @Security("is_granted('ROLE_USER')")
+     */
+    public function handle(Request $request): Response
+    {
+        $id = Uuid::fromString($request->attributes->get('id'))->toString();
+
+        $speaker = $this->speakerRepository->find($id);
+        if (!$speaker) {
+            throw new NotFoundHttpException();
+        }
+
+        return new Response($this->renderer->render('speaker/show.html.twig', [
+            'speaker' => $speaker,
+        ]));
+    }
+}
