+++ src/Speaker/UploadFromOpenCfp/SpeakerTalkCsvUploadRequest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Dto;
+namespace App\Speaker\UploadFromOpenCfp;
 
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\Validator\Constraints as Assert;
