+++ src/Speaker/UploadFromOpenCfp/UploadSpeakerFromCsvFormType.php
@@ -2,15 +2,14 @@
 
 declare(strict_types=1);
 
-namespace App\Form;
+namespace App\Speaker\UploadFromOpenCfp;
 
-use App\Dto\SpeakerTalkCsvUploadRequest;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\FileType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-final class UploadSpeakerFromCsvType extends AbstractType
+final class UploadSpeakerFromCsvFormType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
